Break up `test_management.py` into module
