+++ tests/conftest.py
@@ -6,6 +6,7 @@ from dateutil.relativedelta import relativedelta
 from wagtail.models import Page
 from wagtail.models.i18n import Locale
 
+from caching.private_api.management import CacheManagement
 from cms.home.models.home_page import UKHSARootPage
 from metrics.domain.models import (
     PlotGenerationData,
@@ -349,3 +350,8 @@ def example_trend_number_block() -> dict[str, str]:
 @pytest.fixture
 def timestamp_2_months_from_now() -> datetime.datetime:
     return datetime.datetime.now() + relativedelta(months=2)
+
+
+@pytest.fixture
+def cache_management_with_in_memory_cache() -> CacheManagement:
+    return CacheManagement(in_memory=True)
